webapp/requirement.txt
pandas==1.3.5
jieba==0.42.1
pymysql==0.10.1
pyLDAvis==3.2.2
gensim==4.2.0
PyYAML==3.13
Django ==3.1.7

psutil==5.8.0
django-redis==5.0.0
redis==3.5.3
